requirements.txt
annotated-types==0.6.0
anthropic==0.25.8
anyio==4.3.0
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
netifaces==0.11.0
openai==1.26.0
pickle==3.8
playwright==1.43.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pydantic==2.7.1
pydantic_core==2.18.2
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
beautifulsoup4==4.12.3
matplotlib==3.8.4
numpy==1.26.4
rake-nltk==1.0.6
scikit_learn==1.4.2
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
